video_stream_gdnative: fix delta compensation
The time delta compensation is only needed when seek() isn't
performed in the case of netsync. When seek() is done (time difference
is too big) the time has been updated in ffmpeg videodecoder and we
don't have to set it again in the next update().

This fixes previous commit which causes frame to drop 100% when syncing
to master video playback.